Changes to be committed: 	renamed:    extra/Camera.class -> Camera.class 	renamed:    extra/Camera.java -> Camera.java 	modified:   Environment$1$1.class 	modified:   Environment$1$10.class 	renamed:    extra/Environment$1$11.class -> Environment$1$11.class 	renamed:    extra/Environment$1$12.class -> Environment$1$12.class 	renamed:    extra/Environment$1$13.class -> Environment$1$13.class 	renamed:    extra/Environment$1$14.class -> Environment$1$14.class 	modified:   Environment$1$2.class 	modified:   Environment$1$3.class 	modified:   Environment$1$4.class 	modified:   Environment$1$5.class 	modified:   Environment$1$6.class 	modified:   Environment$1$7.class 	modified:   Environment$1$8.class 	modified:   Environment$1$9.class 	modified:   Environment$1.class 	modified:   Environment.class 	modified:   Environment.java 	renamed:    extra/Map.class -> Map.class 	renamed:    extra/Map.java -> Map.java 	deleted:    Matrix3.class 	deleted:    Matrix3.java 	modified:   Noise.class 	modified:   Noise.java 	renamed:    extra/Player.class -> Player.class 	renamed:    extra/Player.java -> Player.java 	modified:   Renderer.class 	modified:   Renderer.java 	modified:   Triangle.class 	modified:   Triangle.java 	renamed:    extra/Vector3f.class -> Vector3f.class 	renamed:    extra/Vector3f.java -> Vector3f.java 	modified:   Vertex.class 	modified:   Vertex.java 	deleted:    extra.zip 	deleted:    extra/Environment$1$1.class 	deleted:    extra/Environment$1$10.class 	deleted:    extra/Environment$1$2.class 	deleted:    extra/Environment$1$3.class 	deleted:    extra/Environment$1$4.class 	deleted:    extra/Environment$1$5.class 	deleted:    extra/Environment$1$6.class 	deleted:    extra/Environment$1$7.class 	deleted:    extra/Environment$1$8.class 	deleted:    extra/Environment$1$9.class 	deleted:    extra/Environment$1.class 	deleted:    extra/Environment.class 	deleted:    extra/Environment.java 	deleted:    extra/Main.class 	deleted:    extra/Main.java 	deleted:    extra/Noise.class 	deleted:    extra/Noise.java 	deleted:    extra/Renderer.class 	deleted:    extra/Renderer.java 	deleted:    extra/Triangle.class 	deleted:    extra/Triangle.java 	deleted:    extra/Vertex.class 	deleted:    extra/Vertex.java 	renamed:    extra/notes.txt -> notes.txt Seperated the 2 game versions into seperate folders and made the player version the main project

+++ Vertex.java
@@ -1,35 +1,35 @@
 public class Vertex {
-    private double x;
-    private double y;
-    private double z;
+    private float x;
+    private float y;
+    private float z;
 
-    public Vertex(double x, double y, double z) {
+    public Vertex(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public double getX() {
+    public float getX() {
 		return this.x;
     }
 
-    public double getY() {
+    public float getY() {
         return this.y;
     }
 
-    public double getZ() {
+    public float getZ() {
         return this.z;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public void setZ(double z) {
+    public void setZ(float z) {
         this.z = z;
     }
 }
